test(esm): use float64 position/charge tensors (match gradwave RDTYPE)

+++ tests/unit/test_esm_correction.py
@@ -35,8 +35,8 @@ def _elec(grid, centers, charges, width=0.9):
 
 def test_gaussian_ion_integrates_to_Z():
     grid = _grid(20.0, 100)
-    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]])
-    z = torch.tensor([4.0, 6.0])
+    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]], dtype=torch.float64)
+    z = torch.tensor([4.0, 6.0], dtype=torch.float64)
     rho_ion = gaussian_ion_density(pos, z, grid, 0.8)
     dvol = grid.volume / rho_ion.numel()
     assert abs(float(rho_ion.sum() * dvol) - float(z.sum())) < 1e-6
@@ -46,8 +46,8 @@ def test_potential_is_functional_derivative_of_energy():
     """esm_potential must equal δ(esm_energy)/δρ (finite difference) — this is
     what keeps the SCF variational and Hellmann-Feynman forces consistent."""
     grid = _grid(20.0, 100)
-    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]])
-    z = torch.tensor([4.0, 6.0])
+    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]], dtype=torch.float64)
+    z = torch.tensor([4.0, 6.0], dtype=torch.float64)
     rho = _elec(grid, [[3, 3, 8.5], [3, 3, 11.5]], [4.0, 6.0])
     dv = esm_potential(rho, pos, z, grid)
     dvol = grid.volume / rho.numel()
@@ -68,8 +68,8 @@ def test_potential_is_functional_derivative_of_energy():
 def test_force_matches_finite_difference():
     """Autograd δΔE/δR (the ESM force term) matches a finite difference of ΔE."""
     grid = _grid(20.0, 100)
-    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]])
-    z = torch.tensor([4.0, 6.0])
+    pos = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]], dtype=torch.float64)
+    z = torch.tensor([4.0, 6.0], dtype=torch.float64)
     rho = _elec(grid, [[3, 3, 8.5], [3, 3, 11.5]], [4.0, 6.0])
     p = pos.clone().requires_grad_(True)
     (fgrad,) = torch.autograd.grad(esm_energy(rho, p, z, grid), p)
@@ -86,8 +86,8 @@ def test_energy_independent_of_ion_width():
     """The discretization-matched difference makes ΔE independent of the Gaussian
     ion width β (the short-range self-field cancels open-vs-periodic)."""
     grid = _grid(24.0, 120)
-    pos = torch.tensor([[3.0, 3.0, 10.0], [3.0, 3.0, 14.0]])
-    z = torch.tensor([4.0, 6.0])
+    pos = torch.tensor([[3.0, 3.0, 10.0], [3.0, 3.0, 14.0]], dtype=torch.float64)
+    z = torch.tensor([4.0, 6.0], dtype=torch.float64)
     rho = _elec(grid, [[3, 3, 11], [3, 3, 13]], [4.0, 6.0])
     des = [float(esm_energy(rho, pos, z, grid, beta=b)) for b in (0.2, 0.4, 0.6)]
     assert max(des) - min(des) < 1e-3, des
@@ -96,9 +96,9 @@ def test_energy_independent_of_ion_width():
 def test_total_energy_box_independent():
     """E_open = E_periodic + ΔE is independent of vacuum thickness while the
     periodic part alone drifts — the surface-energetics payoff, at fixed dz."""
-    ions = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]])
+    ions = torch.tensor([[3.0, 3.0, 8.0], [3.0, 3.0, 12.0]], dtype=torch.float64)
     elec = [[3, 3, 9.0], [3, 3, 11.0]]  # charge-transfer dipole
-    z = torch.tensor([4.0, 6.0])
+    z = torch.tensor([4.0, 6.0], dtype=torch.float64)
 
     def eo_ep(az, n_open):
         grid = _grid(az, n_open)
